optimize reactor

+++ src/main/kotlin/jeremynoesen/bonebot/modules/Command.kt
@@ -69,14 +69,12 @@ object Command {
             val msg = message.contentRaw.toLowerCase()
             if (msg.startsWith(commandPrefix)) {
                 for (command in commands.keys) {
-                    if (msg.startsWith(command)) {
-                        if ((System.currentTimeMillis() - prevTime) >= cooldown * 1000) {
-                            prevTime = System.currentTimeMillis()
-                            message.channel.sendMessage(
-                                commands[command]!!.replace("\$USER$", message.author.asMention)).queue()
-                        } else {
-                            message.delete().queue()
-                        }
+                    if (msg.startsWith(command) && (System.currentTimeMillis() - prevTime) >= cooldown * 1000) {
+                        prevTime = System.currentTimeMillis()
+                        message.channel.sendMessage(commands[command]!!.replace("\$USER$", message.author.asMention))
+                            .queue()
+                    } else {
+                        message.delete().queue()
                     }
                 }
             }
